requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
djangorestframework==3.15.2
drf-yasg==1.21.8
frozenlist==1.5.0
geoip2==4.8.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
maxminddb==2.6.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
yarl==1.18.3
